move button to game component
